+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### 🚀 Features
+
+- Scale now run apply to the server
+- Add caddy to user group during deployment
+
 ## [0.18.0] - 2026-01-18
 
 ### 🚀 Features